feat/network: jsdoc documentation complete
ADD jsdoc documentation on each components
COMPOSE SENS factory class which generate SENS services

TODO - Changelog , Readme documentation
