+++ src/types/auth_types.ts
@@ -7,4 +7,36 @@
  export type NCPAuthKeyType = {
   accessKey: string
   secretKey: string
+ }
+
+/**
+ * SMSserviceAuthType
+ * @memberof SMS
+ * @alias SMSserviceAuthType
+ */
+export type SMSserviceAuthType = {
+  phone:      string
+  serviceId: string
 }
+
+/**
+ * ApiSignatureParamType
+ * @memberof generateApiSignature
+ * @alias ApiSignatureParamType
+ */
+ export type ApiSignatureParamType = {
+  method: string
+  url: string
+  ncpAuthKey: NCPAuthKeyType
+}
+
+/**
+ * ApiSignatureReturnType
+ * @memberof generateApiSignature
+ * @alias ApiSignatureReturnType
+ */
+ export type ApiSignatureReturnType = {
+  timestamp: string
+  signature: string
+}
+
